refactor(editor): move focusedField listening and setFocusField to lib
and scope anchorIdPrefix in component
