+++ apps/metadata-editor/src/app/edit/edit-page.component.html
@@ -74,7 +74,6 @@
           <gn-ui-metadata-quality-panel
             [record]="facade.record$ | async"
             [editorConfig]="facade.editorConfig$ | async"
-            (criterionClicked)="onCriterionClicked($event)"
           >
           </gn-ui-metadata-quality-panel>
         }
