+++ libs/feature/editor/src/lib/models/editor-config.model.ts
@@ -1,6 +1,5 @@
 import { CatalogRecordKeys } from '@geonetwork-ui/common/domain/model/record'
 
-export const ANCHOR_ID_PREFIX = 'gn-ui--field-'
 // Expressions should be enclosed in `${}` to be recognized as such
 // eg. ${dateNow()}
 export type EditorFieldExpression = `$\{${string}}`
